Several small changes from code review
