+++ src/lib/y2storage/dialogs/guided_setup/base.rb
@@ -93,9 +93,8 @@ module Y2Storage
           true
         end
 
-        # Should be implemented by derived classes.
+        # To be implemented by derived classes, if needed
         def initialize_widgets
-          log.error("ERROR: initialize_widgets not implemented by derived class")
           nil
         end
