+++ test/y2storage/dialogs/guided_setup/select_filesystem/legacy_test.rb
@@ -20,7 +20,7 @@
 # To contact SUSE LLC about this file by physical or electronic mail, you may
 # find current contact information at www.suse.com.
 
-require_relative "../../spec_helper.rb"
+require_relative "../../../spec_helper.rb"
 require_relative "#{TEST_PATH}/support/guided_setup_context"
 
 describe Y2Storage::Dialogs::GuidedSetup::SelectFilesystem::Legacy do
