+++ test/y2storage/dialogs/guided_setup/select_filesystem/volume_widget_test.rb
@@ -20,11 +20,10 @@
 # To contact SUSE LLC about this file by physical or electronic mail, you may
 # find current contact information at www.suse.com.
 
-require_relative "../../spec_helper.rb"
-require_relative "#{TEST_PATH}/support/guided_setup_context"
-require_relative "#{TEST_PATH}/support/proposal_context"
+require_relative "../../../spec_helper.rb"
+require "y2storage/dialogs/guided_setup"
 
-describe Y2Storage::Dialogs::GuidedSetup::SelectFilesystem::Ng::VolumeWidget do
+describe Y2Storage::Dialogs::GuidedSetup::SelectFilesystem::VolumeWidget do
   # Convenience method to inspect the tree of terms for the UI
   def term_with_id(regexp, content)
     content.nested_find do |param|
@@ -105,7 +104,7 @@ describe Y2Storage::Dialogs::GuidedSetup::SelectFilesystem::Ng::VolumeWidget do
       end
     end
 
-    context "if the user can not decide whether to enlarge by ram" do
+    context "if the user cannot decide whether to enlarge by ram" do
       let(:index) { 0 }
 
       it "does not include the corresponding check box" do
@@ -518,108 +517,6 @@ describe Y2Storage::Dialogs::GuidedSetup::SelectFilesystem::Ng::VolumeWidget do
           widget.handle(widget_id)
         end
       end
-    end
-  end
-end
-
-describe Y2Storage::Dialogs::GuidedSetup::SelectFilesystem::Ng do
-  include_context "guided setup requirements"
-
-  subject(:dialog) { described_class.new(guided_setup) }
-
-  before { allow(settings).to receive(:volumes).and_return volumes }
-
-  let(:widget_class) { Y2Storage::Dialogs::GuidedSetup::SelectFilesystem::Ng::VolumeWidget }
-
-  describe "#run" do
-    let(:volumes) do
-      [
-        double("VolumeSpecification", configurable?: true),
-        double("VolumeSpecification", configurable?: false),
-        double("VolumeSpecification", configurable?: true)
-      ]
-    end
-
-    let(:widget0) { double("VolumeWidget", content: [], init: true, store: true) }
-    let(:widget2) { double("VolumeWidget", content: [], init: true, store: true) }
-
-    it "shows a VolumeWidget for every configurable volume" do
-      expect(widget_class).to receive(:new).with(settings, 0).and_return widget0
-      expect(widget_class).to receive(:new).with(settings, 2).and_return widget2
-      expect(widget0).to receive(:content)
-      expect(widget0).to receive(:init)
-      expect(widget2).to receive(:content)
-      expect(widget2).to receive(:init)
-
-      dialog.run
-    end
-
-    it "ignores volumes that are not configurable" do
-      expect(widget_class).to_not receive(:new).with(settings, 1)
-      dialog.run
-    end
-
-    it "calls VolumeWidget#store for all the widgets" do
-      allow(widget_class).to receive(:new).and_return(widget0, widget2)
-      expect(widget0).to receive(:store)
-      expect(widget2).to receive(:store)
-
-      dialog.run
-    end
-  end
-
-  describe "skip?" do
-    context "when the settings contain no volumes" do
-      let(:volumes) { [] }
-
-      it "returns true" do
-        expect(dialog.skip?).to eq true
-      end
-    end
-
-    context "when none of the volumes are configurable" do
-      let(:volumes) do
-        [
-          double("VolumeSpecification", configurable?: false),
-          double("VolumeSpecification", configurable?: false)
-        ]
-      end
-
-      it "returns true" do
-        expect(dialog.skip?).to eq true
-      end
-    end
-
-    context "when some volume is configurable" do
-      let(:volumes) do
-        [
-          double("VolumeSpecification", configurable?: false),
-          double("VolumeSpecification", configurable?: true)
-        ]
-      end
-
-      it "returns false" do
-        expect(dialog.skip?).to eq false
-      end
-    end
-  end
-
-  describe "#handle_event" do
-    let(:volumes) do
-      [
-        double("VolumeSpecification", configurable?: true),
-        double("VolumeSpecification", configurable?: true)
-      ]
-    end
-    let(:widget0) { double("VolumeWidget") }
-    let(:widget1) { double("VolumeWidget") }
-    let(:event) { "an event" }
-
-    it "delegates to the #handle method of all volume widgets" do
-      allow(widget_class).to receive(:new).and_return(widget0, widget1)
-      expect(widget0).to receive(:handle).with(event)
-      expect(widget1).to receive(:handle).with(event)
-      dialog.handle_event(event)
     end
   end
 end
